I've fixed the issue with the language selector images not appearing. Here's what I did:
- I imported the logo images directly into the CosyLanguageSelector component.
- I removed the intermediate logos/index.js file.
- I fixed a linting error caused by a duplicate import.
