--- src/assets/icons/cosylanguages_logos/index.js
@@ -1,31 +1,0 @@
-import cosyarmenian from './cosyarmenian.png';
-import cosybachkir from './cosybachkir.png';
-import cosybreton from './cosybreton.png';
-import cosyenglish from './cosyenglish.png';
-import cosyfrench from './cosyfrench.png';
-import cosygeorgian from './cosygeorgian.png';
-import cosygerman from './cosygerman.png';
-import cosygreek from './cosygreek.png';
-import cosyitalian from './cosyitalian.png';
-import cosyportuguese from './cosyportuguese.png';
-import cosyrussian from './cosyrussian.png';
-import cosyspanish from './cosyspanish.png';
-import cosytatar from './cosytatar.png';
-
-const logos = {
-  cosyarmenian,
-  cosybachkir,
-  cosybreton,
-  cosyenglish,
-  cosyfrench,
-  cosygeorgian,
-  cosygerman,
-  cosygreek,
-  cosyitalian,
-  cosyportuguese,
-  cosyrussian,
-  cosyspanish,
-  cosytatar,
-};
-
-export default logos;
